logout button only needed in portal error boundary
