+++ src/web/utils/PortalErrorBoundary.scss
@@ -1,0 +1,7 @@
+.portal-error-boundary {
+  .logout-button {
+    margin-top: 2rem;
+    margin-right: 3rem;
+    float: right;
+  }
+}
